drop -march=native from CXXFLAGS
TL;DR: never ever use -march=native in CXXFLAGS for distro builds.

-march=native optimizes the build for the specific hardware it is
built on; this means that, for example, newer hardware features such
as AVX on x86_64, may be used, rendering the resulting binaries
completely unusable on hardware without the build-time bits used.
Because of this, -march=native is a giant no-no in distribution builds,
as you never want to optimize something based on the specific buildd
used, but rather make sure it will work on the architecture in general,
in a way compatible with the architecture baseline.